requirements.txt
numpy==1.21.2
werkzeug==2.0.3
flask==2.0.1
flask-cors==3.0.10
opencv-python-headless==4.5.4.58
pillow==9.0.0
yt-dlp==2023.3.4
torch==2.0.0
torchvision==0.15.0